refactor(schedule): replace server-side schedule page with client
Replace the large server-rendered schedule page implementation with a
-side ScheduleClient component Remove unused imports and
previous ScheduleDay/getDaySchedule server logic, actions, UI primitives,
and modal components that were part of the old page.

This simplifies the page entry file, shifts rendering and interactive
behavior to ScheduleClient, and prepares the app for client-driven
scheduling UI and state handling.
